deWarning

+++ include/boundarybcm.h
@@ -32,17 +32,17 @@ void boundarybcm(const bool standard, const double h, const size_t low, const si
     }
     std::cout << "dinterfaces size " << " " <<dinterfaces.size() << std::endl;*/
     if(type == 0 || type == 1){	
-        for ( int i = 0 ; i < dblocks.size();i++){
+        for ( int i = 0 ; i < static_cast<int>(dblocks.size());i++){
             START_RANGE(i)
             if(dconductors[i][0].isempty == false){
-                for ( int j = 0 ; j < dconductors[i].size(); j++){
+                for ( int j = 0 ; j < static_cast<int>(dconductors[i].size()); j++){
                     divide_quadrangle(&dconductors[i][j].windows.vertex[0],
                                     &dconductors[i][j].windows.vertex[1],&dconductors[i][j].windows.vertex[2],
                                     &dconductors[i][j].windows.vertex[3],dconductors[i][j].windows.xEle,
                                     dconductors[i][j].windows.yEle,dconductors[i][j].inverse,i,false);
                 }
             }
-            for(int j = 0 ; j < dblocks[i].size() ; j++){
+            for(int j = 0 ; j < static_cast<int>(dblocks[i].size()) ; j++){
                 divide_quadrangle(&dblocks[i][j].windows.vertex[0],
                                 &dblocks[i][j].windows.vertex[1],&dblocks[i][j].windows.vertex[2],
                                 &dblocks[i][j].windows.vertex[3],dblocks[i][j].windows.xEle,
@@ -66,9 +66,9 @@ void boundarybcm(const bool standard, const double h, const size_t low, const si
                         
                     }	
                     
-                }else if ( i < dblocks.size()-1 && i > 0){
+                }else if ( i < static_cast<int>(dblocks.size())-1 && i > 0){
                     //	std::cout << i << "boundary " << std::endl;
-                    for(int k = 0 ; k < BdElementInter.size();k++){
+                    for(int k = 0 ; k < static_cast<int>(BdElementInter.size());k++){
                         inversePlane(BdElementInter[k]);
                         BdElementsbcm.push_back(BdElementInter[k]);
                         cur_ele_id[i]++;
@@ -88,7 +88,7 @@ void boundarybcm(const bool standard, const double h, const size_t low, const si
                     }	
                     
                 }else {
-                    for(int k = 0 ; k < BdElementInter.size();k++){
+                    for(int k = 0 ; k < static_cast<int>(BdElementInter.size());k++){
                         inversePlane(BdElementInter[k]);
                         BdElementsbcm.push_back(BdElementInter[k]);
                         cur_ele_id[i]++;
@@ -111,7 +111,7 @@ void boundarybcm(const bool standard, const double h, const size_t low, const si
             /*std::cout << "inverse _ i " << i << " " << inverse_i << std::endl;*/
             START_RANGE(inverse_i)
             if(dconductors[i][0].isempty == false){
-                for ( int j = 0 ; j < dconductors[i].size(); j++){
+                for ( int j = 0 ; j < static_cast<int>(dconductors[i].size()); j++){
                     divide_quadrangle(&dconductors[i][j].windows.vertex[0],
                                     &dconductors[i][j].windows.vertex[1],&dconductors[i][j].windows.vertex[2],
                                     &dconductors[i][j].windows.vertex[3],dconductors[i][j].windows.xEle,
@@ -120,23 +120,23 @@ void boundarybcm(const bool standard, const double h, const size_t low, const si
                 }
                 
             }
-            for(int j = 0 ; j < dblocks[i].size() ; j++){
+            for(int j = 0 ; j < static_cast<int>(dblocks[i].size()) ; j++){
                 divide_quadrangle(&dblocks[i][j].windows.vertex[0],&dblocks[i][j].windows.vertex[1],
                                 &dblocks[i][j].windows.vertex[2],&dblocks[i][j].windows.vertex[3],
                                 dblocks[i][j].windows.xEle,dblocks[i][j].windows.yEle,dblocks[i][j].inverse,inverse_i,false);
                 
             }
             if( i == 0){
-                for(int k = 0 ; k < BdElementInter.size();k++){
+                for(int k = 0 ; k < static_cast<int>(BdElementInter.size());k++){
                     inversePlane(BdElementInter[k]);
                     BdElementsbcm.push_back(BdElementInter[k]);
                     cur_ele_id[inverse_i]++;
                     
                 }
                 
-            }else if ( i < dblocks.size()-1 && i > 0){
+            }else if ( i < static_cast<int>(dblocks.size())-1 && i > 0){
                 std::vector<TBDELEINFO> TempBdElementInter;
-                for(int k = 0 ; k < BdElementInter.size();k++){
+                for(int k = 0 ; k < static_cast<int>(BdElementInter.size());k++){
                     inversePlane(BdElementInter[k]);
                     TempBdElementInter.push_back(BdElementInter[k]);
                     
@@ -152,7 +152,7 @@ void boundarybcm(const bool standard, const double h, const size_t low, const si
                                     dinterfaces[i].windows.yEle,false,inverse_i,true);
                     
                 }	
-                for(int k = 0 ; k < BdElementInter.size();k++){
+                for(int k = 0 ; k < static_cast<int>(BdElementInter.size());k++){
                     inversePlane(BdElementInter[k]);
                     BdElementsbcm.push_back(BdElementInter[k]);
                     cur_ele_id[inverse_i]++;
