+++ src/y.tab.cpp
@@ -1755,7 +1755,7 @@ case 38:
 			if(frw){
 				pddlInterfacePoint3DNum = pddlPoint3DNum;
 				pddlInterfacePoint3DBuf = new TPOINT[pddlInterfacePoint3DNum];
-           			 for(int i = 0 ; i < pddlPoint3DNum ; i++) 
+           			 for(int i = 0 ; i < static_cast<int>(pddlPoint3DNum) ; i++) 
 					pddlInterfacePoint3DBuf[i] = pddlPoint3DBuf[i]; 
          				   pddlPoint3DNum = 0; 
 			}if(bcm){ // add by [zhangbl]
@@ -1764,7 +1764,7 @@ case 38:
 				for (int i = 0 ; i < Curblk-1 ; i++){
 					pddlInterfacePoint3DBufMul[i] = new TPOINT [pddlInterfacePoint3DNum];
 				}
-				for ( int i = 0 ; i < pddlPoint3DNum ; i ++){
+				for ( int i = 0 ; i < static_cast<int>(pddlPoint3DNum) ; i ++){
 					pddlInterfacePoint3DBufMul[InterfaceNum][i]= pddlPoint3DBuf[i];
 					pddlPoint3DNum = 0 ;
 					 
@@ -1838,7 +1838,7 @@ case 50:
 {
 		   pddlOutLoopPoint2DNum = pddlPoint2DNum;
 		   int i;
-		   for(i = 0 ; i < pddlPoint2DNum ; i++) 
+		   for(i = 0 ; i < static_cast<int>(pddlPoint2DNum) ; i++) 
 			  pddlOutLoopPoint2DBuf[i] = pddlPoint2DBuf[i]; 
 		   pddlPoint2DNum = 0; 
 	   ;
@@ -1849,7 +1849,7 @@ case 51:
 		//add by [whquan]
             pddlOutLoopPoint3DNum = pddlPoint3DNum;
 			int i;
-            for( i = 0 ; i < pddlPoint3DNum ; i++) 
+            for( i = 0 ; i < static_cast<int>(pddlPoint3DNum) ; i++) 
 				pddlOutLoopPoint3DBuf[i] = pddlPoint3DBuf[i]; 
             pddlPoint3DNum = 0; 
 	   ;
@@ -2206,7 +2206,7 @@ int DoAlpha(ifstream *input)
 		return 300+i;
 	}
 
-   for(i=0;i<ObjNum;i++)
+   for(i=0;i<static_cast<int>(ObjNum);i++)
       if(strcmp(str,ObjField[i].Name)==0)
 	{
 		yylval.ival=i;
